resistance function as module file
